requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
packaging==20.3
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
vulture==1.4
wcwidth==0.1.9
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
zipp==3.1.0
